@@ -11,9 +11,9 @@
     "license": "MIT",
     "authors": [
         {
-            "name": "Max Leonov",
+            "name": "Maksym Leonov",
             "email": "hi@maxleonov.pw",
-            "homepage": "https://maxleonov.pw"
+            "homepage": "https://github.com/maxbanton"
         }
     ],
     "support": {
@@ -21,13 +21,13 @@
         "source": "https://github.com/maxbanton/cwh"
     },
     "require": {
-        "php": "^7.2 || ^8",
-        "monolog/monolog": "^2.0 || ^3.0",
-        "aws/aws-sdk-php": "^3.18"
+        "php": ">=8.1",
+        "monolog/monolog": "^3.0",
+        "aws/aws-sdk-php": "^3.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5 || ^9.4",
-        "squizlabs/php_codesniffer": "^3.5"
+        "phpunit/phpunit": "^9.5",
+        "squizlabs/php_codesniffer": "^3.7"
     },
     "suggest": {
         "maxbanton/dd": "Minimalistic dump-and-die function for easy debugging"
